adAi/requirements.txt
fastapi
pillow
torch
diffusers
uvicorn
